Working 24th March, v1.1.0 uploaded

+++ LCSFinder.cpp
@@ -12,22 +12,22 @@ using namespace std;
 #include "LCSFinder.h"
 
 
-LCSFinder::LCSFinder(const vector<int32_t>& s1, const vector<int32_t>& s2) {
+LCSFinder::LCSFinder(const vector<long long int>& s1, const vector<long long int>& s2) {
     s1sz = s1.size();
     s2sz = s2.size();
     s.insert(s.end(), s1.begin(), s1.end());
-    // Assumes int32_t min is never used in s1 and s2 so it can be used a special character
+    // Assumes long long int min is never used in s1 and s2 so it can be used a special character
     // which is < all other characters
-    s.push_back(numeric_limits<int32_t>::min());
+    s.push_back(numeric_limits<long long int>::min());
     s.insert(s.end(), s2.begin(), s2.end());
 
     BuildSuffixArray();
 }
 
-int32_t LCSFinder::FindLCP(int32_t i, int32_t j) {
+long long int LCSFinder::FindLCP(long long int i, long long int j) {
     // Catches cases when precomp wasn't done or empty strings
     if (prank.size() < 1) return -1;
-    int32_t pid = prank.size()-1, p2 = 1<<pid, ans = 0;
+    long long int pid = prank.size()-1, p2 = 1<<pid, ans = 0;
     // Essentially do binary counting on matching prefix size using prank
     while (pid >= 0 && i < s.size() && j < s.size()) {
         if (i+p2 <= s.size() && j+p2 <= s.size() && prank[pid][i] == prank[pid][j]) {
@@ -43,32 +43,32 @@ int32_t LCSFinder::FindLCP(int32_t i, int32_t j) {
 
 
 
-vector<int32_t> LCSFinder::ComputeAllLCSs(vector<pair<int32_t, int32_t>>& inds) {
-    // Computes all the Longest Common Substring (LCS) values for each pair of indexes int32_to s1 and s2.
-    // The first part of a pair is an index int32_to s1 and the second is int32_to s2.
+vector<long long int> LCSFinder::ComputeAllLCSs(vector<pair<long long int, long long int>>& inds) {
+    // Computes all the Longest Common Substring (LCS) values for each pair of indexes long long into s1 and s2.
+    // The first part of a pair is an index long long into s1 and the second is long long into s2.
     // It is assumed inds is sorted.
     // It is assumed that both the first and second element of each pair are monotonically increasing.
-    // Since each is a "time point32_t", the indexes int32_to s1 and s2 should increase monotonically.
+    // Since each is a "time polong long int", the indexes long long into s1 and s2 should increase monotonically.
     //
     // Returns a list of all the LCS values.
-    // For a given pair, say that the first index int32_to s1 is called i and the second int32_to s2 is called j.
+    // For a given pair, say that the first index long long into s1 is called i and the second long long into s2 is called j.
     // The corresponding entry in the return list is the LCS in prefix s2[0..j) that matches a prefix of the string s1[i..n).
-    vector<int32_t> ans, sa_loc(s.size());
-    for (int32_t i = 0; i < s.size(); ++i) sa_loc[suff_arr[i]] = i;
-    set<int32_t> active;
-    int32_t idx = 0;
+    vector<long long int> ans, sa_loc(s.size());
+    for (long long int i = 0; i < s.size(); ++i) sa_loc[suff_arr[i]] = i;
+    set<long long int> active;
+    long long int idx = 0;
     for (auto& ind : inds) {
-        int32_t s1idx = ind.first;
-        int32_t s2idx = s1sz+1+idx;
+        long long int s1idx = ind.first;
+        long long int s2idx = s1sz+1+idx;
         while (idx < ind.second) {
             active.insert(sa_loc[s2idx]);
             ++idx;
             s2idx = s1sz+1+idx;
         }
-        int32_t lcs = 0;
+        long long int lcs = 0;
         auto it = active.lower_bound(sa_loc[s1idx]);
         while (it != active.end()) {
-            int32_t match = FindLCP(suff_arr[*it], s1idx);
+            long long int match = FindLCP(suff_arr[*it], s1idx);
             if (suff_arr[*it]+match > s2idx) {
                 match = s2idx-suff_arr[*it];
             } else {
@@ -81,7 +81,7 @@ vector<int32_t> LCSFinder::ComputeAllLCSs(vector<pair<int32_t, int32_t>>& inds) 
         it = active.lower_bound(sa_loc[s1idx]);
         while (it != active.begin()) {
             it = prev(it);
-            int32_t match = FindLCP(suff_arr[*it], s1idx);
+            long long int match = FindLCP(suff_arr[*it], s1idx);
             if (suff_arr[*it]+match > s2idx) {
                 match = s2idx-suff_arr[*it];
             } else {
@@ -95,10 +95,10 @@ vector<int32_t> LCSFinder::ComputeAllLCSs(vector<pair<int32_t, int32_t>>& inds) 
     return ans;
 }
 
-vector<int32_t> LCSFinder::GetS() {
+vector<long long int> LCSFinder::GetS() {
     return s;
 }
-vector<int32_t> LCSFinder::GetSA() {
+vector<long long int> LCSFinder::GetSA() {
     return suff_arr;
 }
 
@@ -110,25 +110,25 @@ void LCSFinder::BuildSuffixArray() {
     auto alphabet = s;
     sort(alphabet.begin(), alphabet.end());
     alphabet.erase(unique(alphabet.begin(), alphabet.end()), alphabet.end());
-    for (int32_t& e : s)
+    for (long long int& e : s)
         e = lower_bound(alphabet.begin(), alphabet.end(), e) - alphabet.begin();
 
-    vector<int32_t> rank(s.size()), tmp_rank(s.size());
+    vector<long long int> rank(s.size()), tmp_rank(s.size());
     prank.clear();
     // Sort by power of 2 length prefixes
-    for (int32_t i = 0; i < s.size(); ++i) rank[i] = s[i];
-    for (int32_t p2 = 1; p2 <= s.size(); p2 *= 2) {
+    for (long long int i = 0; i < s.size(); ++i) rank[i] = s[i];
+    for (long long int p2 = 1; p2 <= s.size(); p2 *= 2) {
         prank.push_back(rank);
-        sort(begin(suff_arr), end(suff_arr), [&](int32_t i, int32_t j) {
-            int32_t r2i = i+p2 < s.size() ? rank[i+p2] : -1;
-            int32_t r2j = j+p2 < s.size() ? rank[j+p2] : -1;
+        sort(begin(suff_arr), end(suff_arr), [&](long long int i, long long int j) {
+            long long int r2i = i+p2 < s.size() ? rank[i+p2] : -1;
+            long long int r2j = j+p2 < s.size() ? rank[j+p2] : -1;
             return rank[i] == rank[j] ? r2i < r2j : rank[i] < rank[j];
         });
-        int32_t ridx = 0;
+        long long int ridx = 0;
         tmp_rank[suff_arr[0]] = 0;
-        for (int32_t i = 1; i < s.size(); ++i) {
-            int32_t r2 = suff_arr[i]+p2 < s.size() ? rank[suff_arr[i]+p2] : -1;
-            int32_t r2prev = suff_arr[i-1]+p2 < s.size() ? rank[suff_arr[i-1]+p2] : -1;
+        for (long long int i = 1; i < s.size(); ++i) {
+            long long int r2 = suff_arr[i]+p2 < s.size() ? rank[suff_arr[i]+p2] : -1;
+            long long int r2prev = suff_arr[i-1]+p2 < s.size() ? rank[suff_arr[i-1]+p2] : -1;
             if (rank[suff_arr[i]] != rank[suff_arr[i-1]] || r2 != r2prev)
                 ridx++;
             tmp_rank[suff_arr[i]] = ridx;
